Enable testing of resolving overdue tasks
This is a huge commit, that looks to set up the
prototype as if some tasks are complete and two
are overdue. This should let us show a procurement
operations person how they might resolve this.

There is a common change date page that allows
users to set a new due date and resolve the
overdue status