FIX refine specification with school status tag
The routing expects the same values that the
checkboxes submit. Also added back in the deleted
empty checkbox to ensure that the status is set
to to do if no checkboxes are selected.

This now means that the status tag is always
updated and that the user can always see
what stage they are at. It supports complete,
in progress and to do statuses.